batch function

+++ src/RL_toolbox/reward.py
@@ -299,8 +299,8 @@ class VILBERT(Reward):
         vil_prediction, target = self.get_preds(question)
         if self.reduced_answers:
             mask = torch.ones_like(vil_prediction) * float("-Inf")
-            mask[:, self.dataset.reduced_answers.squeeze()] = vil_prediction[:,
-                                                              self.dataset.reduced_answers.squeeze()]
+            mask[:, self.dataset.reduced_answers.squeeze().long()] = vil_prediction[:,
+                                                              self.dataset.reduced_answers.squeeze().long()]
             vil_prediction = mask
         sorted_logits, sorted_indices = torch.sort(vil_prediction, descending=True)
         ranks = (sorted_indices.squeeze()[..., None] == (target != 0).nonzero().squeeze()).any(-1).nonzero()
